+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 #### 0.1.1 (2015-11-21)
 * *Breaking change:* For the **Force**, use ```options.minPos``` and ```options.maxPos``` to set ```options.layerWidth``` for the **Distributor** inside it instead of having to specify ```options.layerWidth``` explicitly.
 
+* *Breaking change:* Change ```options.labelHeight``` in **Renderer** to ```options.nodeHeight```.
+
 <a name="0.1.0"></a>
 #### 0.1.0 (2015-11-20)